pokedex/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==4.0
gunicorn==20.1.0
psycopg2-binary~=2.8.4
flask==2.1.1
python-decouple
